деление проекта на модули
